+++ src/util.h
@@ -3,22 +3,22 @@
 #include <random>
 #include <vector>
 
-std::vector<double> rdbl_vec;
+std::vector<float> rdbl_vec;
 
-inline double rdbl(int x = 1) {
+inline float rdbl(int x = 1) {
     if (x == 0) return 0; // lets us have a 'clean' single sample
     #ifdef PRECOMPUTE_RDBL
     thread_local static int c = 0;
     return rdbl_vec[++c % rdbl_vec.size()];
     #else
-    thread_local static std::normal_distribution<double> norm(0, 0.25);
+    thread_local static std::normal_distribution<float> norm(0, 0.25);
     thread_local static std::minstd_rand re;
     return norm(re);
     #endif
 }
 
 inline void seedrdbl() {
-    std::normal_distribution<double> norm(0, 0.25);
+    std::normal_distribution<float> norm(0, 0.25);
     std::minstd_rand re;
     rdbl_vec.resize(100000);
     rdbl_vec[0] = 0.;
@@ -28,6 +28,6 @@ inline void seedrdbl() {
 }
 
 
-inline bool dblequ(const double& x, const double& y, const double& EPS = 0.0001) {
+inline bool dblequ(const float& x, const float& y, const float& EPS = 0.0001) {
     return std::abs(x-y) < EPS;
 }
